fix migration file

+++ database/migrations/2024_11_14_000000_add_type_column_to_failed_one_time_operations_table.php
@@ -25,8 +25,10 @@ class AddTypeColumnToFailedOperationsTable extends Migration
 
     public function down(): void
     {
-        Schema::table($this->name, function (Blueprint $table) {
-            $table->dropColumn('type');
-        });
+        if (Schema::hasTable($this->name) && Schema::hasColumn($this->name, 'type')) {
+            Schema::table($this->name, function (Blueprint $table) {
+                $table->dropColumn('type');
+            });
+        }
     }
 }
